fix generate tests: migrate PGType/BaseType from string to typeinfo.Type
Update all test files in internal/generate/ to use typeinfo.MustParse()
for PGType and BaseType struct fields. Domain-backed columns (slug,
email) use typeinfo.Type{Base, DomainName} directly.

Golden file simple_expected.sql updated for normalized type names
(bigint->int8). Expected string in TestMinimalTable updated (integer->int4).

Claude-Code-Session-Id: df7f98a7-e6d9-4c23-aae2-b5db6ec5707d

+++ internal/generate/testdata/simple_expected.sql
@@ -16,7 +16,7 @@ CREATE TABLE shop.customers (
 CREATE TABLE shop.orders (
     id uuid NOT NULL DEFAULT gen_random_uuid(),
     customer_id uuid NOT NULL,
-    total bigint NOT NULL DEFAULT 0,
+    total int8 NOT NULL DEFAULT 0,
     created_at timestamptz NOT NULL DEFAULT now(),
     CONSTRAINT pk_orders PRIMARY KEY (id)
 );
